+++ tilings/strategies/fusion.py
@@ -26,7 +26,6 @@ from comb_spec_searcher.strategies import Rule
 from comb_spec_searcher.strategies.constructor import RelianceProfile, SubRecs
 from tilings import GriddedPerm, Tiling
 from tilings.algorithms import ComponentFusion, Fusion
-from tilings.assumptions import TrackingAssumption
 
 __all__ = ["FusionStrategy", "ComponentFusionStrategy"]
 
@@ -525,7 +524,8 @@ class FusionStrategy(Strategy[Tiling, GriddedPerm]):
         algo = self.fusion_algorithm(comb_class)
         child = children[0]
         mapped_assumptions = [
-            TrackingAssumption(gps) for gps in algo.assumptions_fuse_counters
+            ass.__class__(gps)
+            for ass, gps in zip(comb_class.assumptions, algo.assumptions_fuse_counters)
         ]
         return (
             {
@@ -560,9 +560,8 @@ class FusionStrategy(Strategy[Tiling, GriddedPerm]):
     def _fuse_parameter(self, comb_class: Tiling) -> str:
         algo = self.fusion_algorithm(comb_class)
         child = algo.fused_tiling()
-        fuse_assumption = TrackingAssumption(
-            child.forward_map(gp) for gp in algo.new_assumption().gps
-        )
+        ass = algo.new_assumption()
+        fuse_assumption = ass.__class__(child.forward_map(gp) for gp in ass.gps)
         return child.get_parameter(fuse_assumption)
 
     def formal_step(self) -> str:
@@ -614,32 +613,10 @@ class FusionStrategy(Strategy[Tiling, GriddedPerm]):
         return cls(**d)
 
 
-class ComponentFusionConstructor(FusionConstructor):
-    pass
-
-
 class ComponentFusionStrategy(FusionStrategy):
-    def __init__(self, row_idx=None, col_idx=None, tracked: bool = False):
-        assert not tracked, "tracking not implemented for component fusion"
-        super().__init__(row_idx=row_idx, col_idx=col_idx, tracked=False)
-
     def fusion_algorithm(self, tiling: Tiling) -> Fusion:
-        return ComponentFusion(tiling, row_idx=self.row_idx, col_idx=self.col_idx)
-
-    def constructor(
-        self, comb_class: Tiling, children: Optional[Tuple[Tiling, ...]] = None,
-    ) -> ComponentFusionConstructor:
-        if not self.tracked:
-            # constructor only enumerates when tracked.
-            return ComponentFusionConstructor("n", {}, tuple(), tuple(), tuple())
-        # TODO: extra parameters?
-        if children is None:
-            children = self.decomposition_function(comb_class)
-            if children is None:
-                raise StrategyDoesNotApply("Strategy does not apply")
-        return ComponentFusionConstructor(
-            self._fuse_parameter(comb_class),
-            *self.extra_parameters(comb_class, children),
+        return ComponentFusion(
+            tiling, row_idx=self.row_idx, col_idx=self.col_idx, tracked=self.tracked
         )
 
     def formal_step(self) -> str:
@@ -692,7 +669,6 @@ class FusionFactory(StrategyFactory[Tiling]):
 
 class ComponentFusionFactory(StrategyFactory[Tiling]):
     def __init__(self, tracked: bool = False):
-        assert not tracked, "tracking not implemented for component fusion"
         self.tracked = tracked
 
     def __call__(self, comb_class: Tiling, **kwargs) -> Iterator[Rule]:
@@ -700,22 +676,22 @@ class ComponentFusionFactory(StrategyFactory[Tiling]):
             return
         cols, rows = comb_class.dimensions
         for row_idx in range(rows - 1):
-            algo = ComponentFusion(comb_class, row_idx=row_idx)
+            algo = ComponentFusion(comb_class, row_idx=row_idx, tracked=self.tracked)
             if algo.fusable():
                 fused_tiling = algo.fused_tiling()
-                yield ComponentFusionStrategy(row_idx=row_idx)(
+                yield ComponentFusionStrategy(row_idx=row_idx, tracked=self.tracked)(
                     comb_class, (fused_tiling,)
                 )
         for col_idx in range(cols - 1):
-            algo = ComponentFusion(comb_class, col_idx=col_idx)
+            algo = ComponentFusion(comb_class, col_idx=col_idx, tracked=self.tracked)
             if algo.fusable():
                 fused_tiling = algo.fused_tiling()
-                yield ComponentFusionStrategy(col_idx=col_idx)(
+                yield ComponentFusionStrategy(col_idx=col_idx, tracked=self.tracked)(
                     comb_class, (fused_tiling,)
                 )
 
     def __str__(self) -> str:
-        return "component fusion"
+        return f"{'tracked ' if self.tracked else ''}component fusion"
 
     def __repr__(self) -> str:
         return self.__class__.__name__ + "()"